requirements.txt
Flask==0.10.1
Jinja2==2.7.3
nltk==3.1
numpy==1.9.2
pyenchant==1.6.6
scikit-learn==0.17
scipy==0.16.1
sklearn==0.0
